requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dotenv==0.9.9
Flask==3.1.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
logzero==1.7.0
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
smartapi-python==1.5.5
urllib3==2.3.0
websocket==0.2.1
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
